PAGE_INDEX/requirements.txt
streamlit
pypdf
python-docx
pydantic-ai
google-generativeai
python-dotenv
asyncio
